Move tests to dedicated trait
